requirements.txt
tinytuya>=2.0
piper-tts
pycryptodome
